+可以正常运行了哦

+++ App/Model/Farm/CartModel.php
@@ -3,15 +3,14 @@
 
 namespace App\Model\Farm;
 
-
-use function GuzzleHttp\Promise\all;
-
 /**
  * 购物车类
  * @property $id
  * @property $good_id
  * @property $quantity
  * @property $account_id
+ * @property $products
+ * @property $is_overdue
  * Class CartModel
  * @package App\Model\Farm
  */
@@ -23,50 +22,59 @@ class CartModel extends BaseModel
 
     public function products()
     {
-        $this->hasOne(ProductModel::class, null, 'id', 'good_id');
+        return $this->hasOne(ProductModel::class, null, 'good_id', 'id');
     }
 
-    /**
-     * 获取需要的购物车信息
-     * @param int $account_id
-     * @return array|bool|\EasySwoole\ORM\Collection\Collection|\EasySwoole\ORM\Db\Cursor|\EasySwoole\ORM\Db\CursorInterface
-     * @throws \EasySwoole\ORM\Exception\Exception
-     * @throws \Throwable
-     */
     public function getAll(int $account_id)
     {
-        return  $this->where(['account_id'=>$account_id])->order('id', 'DESC')->with(['products'])->all();
+        return  $this->with(['products'])->where(['account_id'=>$account_id])->order('id', 'DESC')->all();
+    }
+
+    public function getCartToOrder(array $cart_ids)
+    {
+        return $this->where('id', $cart_ids, 'IN')->all();
     }
 
     public function createInfo(array $data):CartModel
     {
-        $this->account_id = $data['user_id'];
-        $this->quantity = $data['quantity'];
-        $this->good_id = $data['good_id'];
+        $old_info = $this->where(['good_id'=>$data['product_id']])->get();
 
-        $this->id = $this->save();
+        if ($old_info) {
+            $old_info->quantity = $old_info->quantity + $data['quantity'];
+            $old_info->update();
 
-        return $this;
+            return $old_info;
+        } else {
+            $this->account_id = $data['user_id'];
+            $this->quantity = $data['quantity'];
+            $this->good_id = $data['product_id'];
+            $this->is_overdue = 0;
+
+            $this->id = $this->save();
+            return $this;
+        }
     }
-
 
     public function getInfo(int $cat_id):CartModel
     {
-       return $this->get(['id'=>$cat_id])->with(['products']);
+       return $this->with(['products'])->get($cat_id);
     }
-
 
     public function updateInfo(array $data):CartModel
     {
-        $cart_id = $data['cat_id'];
-        unset($data['out_trade_no']);
-
-        $this->where(['id' => $cart_id])->update($data);
+        $cart_id = $data['cart_id'];
 
         $now_info = $this->getInfo($cart_id);
-        if ($now_info->quantity==0) {
-            $this->destroy(['id' => $cart_id]);
+        if ($data['deal']==0) {
+            $now_info->quantity = $now_info->quantity - ($data['quantity']??1);
+            if ($now_info->quantity==0) {
+                $this->destroy(['id' => $cart_id]);
+            }
+        } else {
+            $now_info->quantity = $now_info->quantity + ($data['quantity']??1);
         }
+
+        $now_info->update();
 
         return $now_info;
     }
